Remove prefix from all classes, reinstate -xs
By removing the need for -xs it simplified the
more commonly used classes, but it also created
the possibility of conflicting with Bootstrap 4.

To avoid conflicts, the prefix 'x' was added to
all classes.

The lack of -xs on some classes, but need for it
on others is an extra thing to remember.

Reinstating -xs means that all classes now use it,
which is more consistent for the user.

It also means we can remove the '.x' prefix from
all classes in the library.